pyproject.toml
[project]
name = "make-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logfire>=3.14.0",
    "pydantic-ai>=0.1.3",
    "pydantic-ai-slim[mcp]>=0.1.3",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "telegramify-markdown>=0.5.1",
    "tiktoken>=0.8.0",
]
authors = [
    { name = "Ferdous", email = "ferdous@hey.com" }
]
keywords = ["investment", "finance", "trading", "agent", "mcp", "model context protocol", "broker", "tastytrade", "wallstreetbets"]
classifiers = [
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
bot = "make_bot.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["make_bot"]
